+++ projects/recalbox-hardware/case/installer.py
@@ -4,12 +4,12 @@ import cases
 # Module per case(s)
 MODULES = {
     "installers.gpi.install": (cases.GPI_V1, cases.GPI_V2, cases.GPI_V3),
-    "installers.nespi4.install": (cases.NESPI4,),
     "installers.nuxii.install": (cases.NUXII,),
     "installers.piboy.install": (cases.PIBOY,),
     "installers.gpi2.install": (cases.GPI2,),
+    "installers.argonone.install": (cases.ARGONONE,),
+    "installers.retroflags.install": (cases.SUPERPI4CASE, cases.NESPI4, cases.NESPI4MANUAL, cases.PISTATION, cases.NESPICASEPLUS, cases.SUPERPICASE, cases.MEGAPICASE),
 }
-
 
 def processHardware(install, case, previousCase):
 
@@ -30,13 +30,14 @@ def processHardware(install, case, previousCase):
     # Call target installer/uninstaller
     for moduleName, caseNames in MODULES.items():
         if case in caseNames:
-            logger.hardlog("Install new case: " + case)
             module = __import__(moduleName, fromlist=["Install"])
             installClass = getattr(module, "Install")
             installer = installClass()
             if install:
+                logger.hardlog("Installing case hardware: " + case)
                 installReboot = installer.InstallHardware(case)
             else:
+                logger.hardlog("Uninstalling case hardware: " + case)
                 installReboot = installer.UninstallHardware(case)
 
     return uninstallReboot | installReboot
@@ -47,13 +48,14 @@ def processSoftware(install, case):
     # Call target installer/uninstaller
     for moduleName, caseNames in MODULES.items():
         if case in caseNames:
-            logger.hardlog("Install new case: " + case)
             module = __import__(moduleName, fromlist=["Install"])
             installClass = getattr(module, "Install")
             installer = installClass()
             if install:
+                logger.hardlog("Installing case software: " + case)
                 case = installer.InstallSoftware(case)
             else:
+                logger.hardlog("Uninstalling case software: " + case)
                 case = installer.UninstallSoftware(case)
 
     return case
@@ -63,7 +65,7 @@ def getInstallScript(case):
 
     for moduleName, caseNames in MODULES.items():
         if case in caseNames:
-            logger.hardlog("Get install picture: " + case)
+            logger.hardlog("Get install script: " + case)
             module = __import__(moduleName, fromlist=["Install"])
             installClass = getattr(module, "Install")
             installer = installClass()